REFACTOR: doing some cleanup

+++ src/main/java/com/welcometodannyland/pizza/entity/Dough.java
@@ -9,7 +9,7 @@ public class Dough {
 
     @Id
     @GeneratedValue()
-    private long doughId;
+    private int doughId;
     private String doughType;
 
     public String getDoughType() {
